Created InfiniteDetailSongList Component
